requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.2.5
Babel==2.6.0
blessings==1.7
blinker==1.4
Brlapi==0.6.7
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
command-not-found==0.3
cryptography==2.3
cupshelpers==1.0
defer==1.0.6
distro==1.3.0
distro-info===0.18ubuntu0.18.10.2
dnspython==1.16.0
entrypoints==0.2.3.post3
httplib2==0.11.3
idna==2.6
isort==4.3.18
joblib==0.13.2
keyring==15.1.0
keyrings.alt==3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.14.0
lazr.uri==1.0.3
lazy-object-proxy==1.4.0
louis==3.6.0
macaroonbakery==1.1.4
Mako==1.0.9
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
netifaces==0.10.4
oauthlib==2.0.6
olefile==0.46
pdoc3==0.6.1
pexpect==4.6.0
Pillow==5.2.0
protobuf==3.0.0
psutil==5.4.6
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.30.1
PyJWT==1.6.4
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.2.1
pyRFC3339==1.1
python-apt==1.7.1
python-debian==0.1.33
python-xlib==0.23
pytz==2018.5
pyxdg==0.25
PyYAML==3.12
reportlab==3.5.6
requests==2.18.4
requests-unixsocket==0.1.5
safeeyes==2.0.9
scapy==2.4.2
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.15.0
six==1.12.0
system-service==0.3
systemd-python==234
typed-ast==1.3.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
wadllib==1.3.3
wrapt==1.11.1
xkit==0.0.0
